market-morning-brief/requirements.txt
# ── AI 分析引擎（三选一，按需安装）──────────────────────────────────
anthropic>=0.40.0         # Claude API（Anthropic）
openai>=1.30.0            # ChatGPT API（OpenAI）
google-generativeai>=0.8.0  # Gemini API（Google）

# ── 调度器 ────────────────────────────────────────────────────────────
APScheduler>=3.10.4

# ── HTTP 请求 ─────────────────────────────────────────────────────────
requests>=2.31.0
httpx>=0.27.0

# ── HTML 解析 ─────────────────────────────────────────────────────────
beautifulsoup4>=4.12.3
lxml>=5.1.0

# ── RSS 解析 ──────────────────────────────────────────────────────────
feedparser>=6.0.11

# ── 金融数据（免费）──────────────────────────────────────────────────
akshare>=1.14.0           # A股/港股/美股，全免费
yfinance>=0.2.36          # Yahoo Finance，免费
baostock>=0.8.9           # A股历史数据，免费注册

# ── 数据处理 ──────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0

# ── 工具 ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.1      # 读取 .env 文件
pytz>=2024.1              # 时区处理
tenacity>=8.3.0           # 重试装饰器
